+++ src/main/java/net/sf/jabref/gui/journals/ManageJournalsPanel.java
@@ -22,6 +22,7 @@ import java.awt.event.ActionListener;
 import java.awt.event.MouseAdapter;
 import java.awt.event.MouseEvent;
 import java.awt.event.MouseListener;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
@@ -35,6 +36,7 @@ import java.nio.file.StandardOpenOption;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.Optional;
 
 import javax.swing.AbstractAction;
 import javax.swing.Action;
@@ -46,7 +48,6 @@ import javax.swing.InputMap;
 import javax.swing.JButton;
 import javax.swing.JComponent;
 import javax.swing.JDialog;
-import javax.swing.JFileChooser;
 import javax.swing.JLabel;
 import javax.swing.JOptionPane;
 import javax.swing.JPanel;
@@ -57,12 +58,15 @@ import javax.swing.JTextField;
 import javax.swing.table.AbstractTableModel;
 
 import net.sf.jabref.Globals;
-import net.sf.jabref.gui.FileDialogs;
+
 import net.sf.jabref.gui.IconTheme;
 import net.sf.jabref.gui.JabRefFrame;
+import net.sf.jabref.gui.NewFileDialogs;
+import net.sf.jabref.gui.actions.BrowseAction;
 import net.sf.jabref.gui.help.HelpAction;
 import net.sf.jabref.gui.keyboard.KeyBinding;
 import net.sf.jabref.gui.net.MonitoredURLDownload;
+
 import net.sf.jabref.logic.help.HelpFile;
 import net.sf.jabref.logic.journals.Abbreviation;
 import net.sf.jabref.logic.journals.JournalAbbreviationLoader;
@@ -186,45 +190,31 @@ class ManageJournalsPanel extends JPanel {
         journalEditPanel = builder2.getPanel();
 
         viewBuiltin.addActionListener(e -> {
-            JTable table = new JTable(JournalAbbreviationsUtil
-                    .getTableModel(Globals.journalAbbreviationLoader
-                            .getRepository(JournalAbbreviationPreferences.fromPreferences(Globals.prefs))
-                            .getAbbreviations()));
+            JTable table = new JTable(JournalAbbreviationsUtil.getTableModel(Globals.journalAbbreviationLoader
+                    .getRepository(JournalAbbreviationPreferences.fromPreferences(Globals.prefs)).getAbbreviations()));
             JScrollPane pane = new JScrollPane(table);
             JOptionPane.showMessageDialog(null, pane, Localization.lang("Journal list preview"),
                     JOptionPane.INFORMATION_MESSAGE);
         });
 
         browseNew.addActionListener(e -> {
-            String name;
-            if (!newNameTf.getText().isEmpty()) {
-                name = FileDialogs.getNewFile(frame, Paths.get(newNameTf.getText()).toFile(), Collections.emptyList(), JFileChooser.SAVE_DIALOG,
-                        false);
-            } else {
-                name = FileDialogs.getNewFile(frame, null, Collections.emptyList(), JFileChooser.SAVE_DIALOG, false);
-            }
-
-            if (name != null) {
-                newNameTf.setText(name);
+            Optional<Path> path = new NewFileDialogs(frame, newNameTf.getText()).saveNewFile();
+            path.ifPresent(fileName -> {
+                newNameTf.setText(fileName.toString());
                 newFile.setSelected(true);
-            }
+            });
+
         });
 
         browseOld.addActionListener(e -> {
-            String name;
-            if (!personalFile.getText().isEmpty()) {
-                name = FileDialogs.getNewFile(frame, Paths.get(personalFile.getText()).toFile(),
-                        Collections.emptyList(), JFileChooser.OPEN_DIALOG, false);
-            } else {
-                name = FileDialogs.getNewFile(frame, null, Collections.emptyList(), JFileChooser.OPEN_DIALOG, false);
-            }
+            Optional<Path> path = new NewFileDialogs(frame, personalFile.getText()).openDlgAndGetSelectedFile();
 
-            if (name != null) {
-                personalFile.setText(name);
+            path.ifPresent(fileName -> {
+                personalFile.setText(fileName.toString());
                 oldFile.setSelected(true);
                 oldFile.setEnabled(true);
                 setupUserTable();
-            }
+            });
         });
 
         ok.addActionListener(e -> {
@@ -420,19 +410,19 @@ class ManageJournalsPanel extends JPanel {
         public void actionPerformed(ActionEvent e) {
             String chosen;
             chosen = JOptionPane.showInputDialog(Localization.lang("Choose the URL to download."), "");
-            if (chosen == null) {
+            if ((chosen == null) || comp.getText().isEmpty()) {
                 return;
             }
             File toFile;
             try {
-                String toName = FileDialogs.getNewFile(frame, new File(System.getProperty("user.home")),
-                        Collections.emptyList(),
-                        JFileChooser.SAVE_DIALOG, false);
-                if (toName == null) {
-                    return;
+
+                Optional<Path> path = new NewFileDialogs(frame, System.getProperty("user.home")).saveNewFile();
+                if (path.isPresent()) {
+                    toFile = new File(path.get().toString());
                 } else {
-                    toFile = new File(toName);
+                    return;
                 }
+
                 URL url = new URL(chosen);
                 MonitoredURLDownload.buildMonitoredDownload(comp, url).downloadToFile(toFile);
                 comp.setText(toFile.getPath());
@@ -440,34 +430,6 @@ class ManageJournalsPanel extends JPanel {
                 JOptionPane.showMessageDialog(null, Localization.lang("Error downloading file '%0'", chosen),
                         Localization.lang("Download failed"), JOptionPane.ERROR_MESSAGE);
                 LOGGER.debug("Error downloading file", ex);
-            }
-        }
-    }
-
-    class BrowseAction extends AbstractAction {
-
-        private final JTextField comp;
-        private final boolean dir;
-
-
-        public BrowseAction(JTextField tc, boolean dir) {
-            super(Localization.lang("Browse"));
-            this.dir = dir;
-            comp = tc;
-        }
-
-        @Override
-        public void actionPerformed(ActionEvent e) {
-            String chosen;
-            if (dir) {
-                chosen = FileDialogs.getNewDir(frame, new File(comp.getText()), Collections.emptyList(),
-                        JFileChooser.OPEN_DIALOG, false);
-            } else {
-                chosen = FileDialogs.getNewFile(frame, new File(comp.getText()), Collections.emptyList(),
-                        JFileChooser.OPEN_DIALOG, false);
-            }
-            if (chosen != null) {
-                comp.setText(Paths.get(chosen).toString());
             }
         }
     }
@@ -577,7 +539,7 @@ class ManageJournalsPanel extends JPanel {
         }
     }
 
-    class ExternalFileEntry {
+    private class ExternalFileEntry {
 
         private JPanel pan;
         private final JTextField tf;
@@ -599,8 +561,7 @@ class ManageJournalsPanel extends JPanel {
 
         private void setupPanel() {
             tf.setEditable(false);
-            BrowseAction browseA = new BrowseAction(tf, false);
-            browse.addActionListener(browseA);
+            browse.addActionListener(BrowseAction.buildForFile(tf));
             DownloadAction da = new DownloadAction(tf);
             download.addActionListener(da);
             FormBuilder builder = FormBuilder.create().layout(new FormLayout(
@@ -623,9 +584,10 @@ class ManageJournalsPanel extends JPanel {
                     JOptionPane.showMessageDialog(null, pane, Localization.lang("Journal list preview"),
                             JOptionPane.INFORMATION_MESSAGE);
                 } catch (FileNotFoundException ex) {
+                    LOGGER.debug("File not found", ex);
+
                     JOptionPane.showMessageDialog(null, Localization.lang("File '%0' not found", tf.getText()),
                             Localization.lang("Error"), JOptionPane.ERROR_MESSAGE);
-                    LOGGER.debug("File not found", ex);
                 }
             });
             clear.addActionListener(e -> {